refactor(exception): move exception files to properly distributed package
