+++ src/main/java/racingcar/exception/carName/InvalidRegisterCarsInputLengthException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.carName;
 
 public class InvalidRegisterCarsInputLengthException extends RuntimeException {
