+++ src/main/java/racingcar/exception/executionNumber/NotNumberException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.executionNumber;
 
 public class NotNumberException extends IllegalArgumentException {
