+++ src/main/java/racingcar/service/VerificationService.java
@@ -7,13 +7,13 @@ import static racingcar.util.RegisterCarNumberConstant.CAR_NAME_MAX_LENGTH;
 import static racingcar.util.RegisterCarNumberConstant.REGISTER_CARS_INPUT_MAX_LENGTH;
 
 import racingcar.exception.EmptyInputException;
-import racingcar.exception.InvalidCharacterException;
-import racingcar.exception.InvalidCarNameLengthException;
+import racingcar.exception.carName.InvalidCharacterException;
+import racingcar.exception.carName.InvalidCarNameLengthException;
 
 import java.util.List;
-import racingcar.exception.InvalidRegisterCarsInputLengthException;
-import racingcar.exception.NotNumberException;
-import racingcar.exception.NumberRangeException;
+import racingcar.exception.carName.InvalidRegisterCarsInputLengthException;
+import racingcar.exception.executionNumber.NotNumberException;
+import racingcar.exception.executionNumber.NumberRangeException;
 
 public class VerificationService {
